Dynamic NPC Checkpoint System
There is the potential to set the game to have unlimited checkpoints, but currently it's set for a maximum of 5. Ultimately we need to also create a separate spawner to load potential start, checkpoints and finish point -> then spawn the npcs at the start and begin the process of defending whatever. the script will fall through and determine how many checkpoints there are within the scene and handle the NPC's current task and move to the CP.. Currently the NPC will noclip through say placed obj's as I haven't added colliders etc yet.